Move parsing to swr, fix UX when claim fails
