+++ packages/arb-token-bridge-ui/src/pages/api/cctp/[type].ts
@@ -6,10 +6,7 @@ import {
   InMemoryCache
 } from '@apollo/client'
 import { NextApiRequest, NextApiResponse } from 'next'
-import { DepositStatus, MergedTransaction } from '../../../state/app/state'
-import { getStandardizedTimestamp } from '../../../state/app/utils'
-import { CommonAddress } from '../../../util/CommonAddressUtils'
-import { ChainId, isNetwork } from '../../../util/networks'
+import { ChainId } from '../../../util/networks'
 
 const subgraphUrl = process.env.NEXT_PUBLIC_CCTP_SUBGRAPH_BASE_URL
 if (!subgraphUrl) {
@@ -69,19 +66,19 @@ export type MessageSent = {
   amount: string
 }
 
-type PendingCCTPTransfer = {
+export type PendingCCTPTransfer = {
   messageSent: MessageSent
 }
 
-type CompletedCCTPTransfer = PendingCCTPTransfer & {
+export type CompletedCCTPTransfer = PendingCCTPTransfer & {
   messageReceived: MessageReceived
 }
 
 export type Response =
   | {
       data: {
-        pending: MergedTransaction[]
-        completed: MergedTransaction[]
+        pending: PendingCCTPTransfer[]
+        completed: CompletedCCTPTransfer[]
       }
       error: null
     }
@@ -92,98 +89,6 @@ export type Response =
       }
       error: string
     }
-
-export type CCTPSupportedChainId =
-  | ChainId.Mainnet
-  | ChainId.Goerli
-  | ChainId.ArbitrumOne
-  | ChainId.ArbitrumGoerli
-
-function getSourceChainIdFromSourceDomain(
-  sourceDomain: ChainDomain,
-  l1ChainId: ChainId
-): CCTPSupportedChainId {
-  const { isTestnet } = isNetwork(l1ChainId)
-
-  // Deposits
-  if (sourceDomain === ChainDomain.Mainnet) {
-    return isTestnet ? ChainId.Goerli : ChainId.Mainnet
-  }
-
-  // Withdrawals
-  return isTestnet ? ChainId.ArbitrumGoerli : ChainId.ArbitrumOne
-}
-
-export function getUsdcTokenAddressFromSourceChainId(
-  sourceChainId: CCTPSupportedChainId
-) {
-  return {
-    [ChainId.Mainnet]: CommonAddress.Mainnet.USDC,
-    [ChainId.ArbitrumOne]: CommonAddress.ArbitrumOne.USDC,
-    [ChainId.Goerli]: CommonAddress.Goerli.USDC,
-    [ChainId.ArbitrumGoerli]: CommonAddress.ArbitrumGoerli.USDC
-  }[sourceChainId]
-}
-function parseTransferToMergedTransaction(
-  transfer: PendingCCTPTransfer | CompletedCCTPTransfer,
-  l1ChainId: ChainId,
-  isPending: boolean
-): MergedTransaction {
-  const depositStatus = isPending
-    ? DepositStatus.CCTP_SOURCE_SUCCESS
-    : DepositStatus.CCTP_DESTINATION_SUCCESS
-
-  const { messageSent } = transfer
-  let status = 'Unconfirmed'
-  let resolvedAt = null
-  let receiveMessageTransactionHash = null
-  let receiveMessageTimestamp = null
-
-  if ('messageReceived' in transfer) {
-    const { messageReceived } = transfer
-    status = 'Executed'
-    resolvedAt = getStandardizedTimestamp(
-      (parseInt(messageReceived.blockTimestamp, 10) * 1_000).toString()
-    )
-    receiveMessageTransactionHash = messageReceived.transactionHash
-    receiveMessageTimestamp = getStandardizedTimestamp(
-      messageReceived.blockTimestamp
-    )
-  }
-  const sourceChainId = getSourceChainIdFromSourceDomain(
-    parseInt(messageSent.sourceDomain, 10),
-    l1ChainId
-  )
-  const isDeposit =
-    parseInt(messageSent.sourceDomain, 10) === ChainDomain.Mainnet
-
-  return {
-    sender: messageSent.sender,
-    destination: messageSent.recipient,
-    direction: isDeposit ? 'deposit' : 'withdraw',
-    status,
-    createdAt: getStandardizedTimestamp(
-      (parseInt(messageSent.blockTimestamp, 10) * 1_000).toString()
-    ),
-    resolvedAt,
-    txId: messageSent.transactionHash,
-    asset: 'USDC',
-    value: messageSent.amount,
-    uniqueId: null,
-    isWithdrawal: !isDeposit,
-    blockNum: parseInt(messageSent.blockNumber, 10),
-    tokenAddress: getUsdcTokenAddressFromSourceChainId(sourceChainId),
-    depositStatus,
-    isCctp: true,
-    cctpData: {
-      sourceChainId,
-      attestationHash: messageSent.attestationHash,
-      messageBytes: messageSent.message,
-      receiveMessageTransactionHash,
-      receiveMessageTimestamp
-    }
-  }
-}
 
 export default async function handler(
   req: NextApiRequestWithCCTPParams,
@@ -337,23 +242,20 @@ export default async function handler(
         // If the MessageSent has a corresponding MessageReceived
         const messageReceived = messagesReceivedMap.get(messageSent.id)
         if (messageReceived) {
-          acc.completed.push(
-            parseTransferToMergedTransaction(
-              { messageSent, messageReceived },
-              l1ChainId,
-              false
-            )
-          )
+          acc.completed.push({
+            messageReceived,
+            messageSent
+          })
         } else {
-          acc.pending.push(
-            parseTransferToMergedTransaction({ messageSent }, l1ChainId, true)
-          )
+          acc.pending.push({
+            messageSent
+          })
         }
         return acc
       },
       { completed: [], pending: [] } as {
-        completed: MergedTransaction[]
-        pending: MergedTransaction[]
+        completed: CompletedCCTPTransfer[]
+        pending: PendingCCTPTransfer[]
       }
     )
